Create RA_Demo.py
Beginnings of a demonstration which uses a preliminary dataset and YOLOv8 model (converted to a .blob format) deployed on the Luxonis OAK-D W camera to detect the Dobot Magician Robot Arm and End Effector as well as draw bounding boxes and label depth and coordinates. Currently, the labelling is incorrect.